fix: replaced changes to icon and removed fd-icon according to review
