--- src/fd-icon.scss
@@ -1,20 +1,0 @@
-@import "./mixins";
-
-$block: #{$fd-namespace}-icon;
-
-.#{$block} {
-  @include fd-reset();
-  @include fd-icon-base();
-
-  font-size: 1rem;
-  font-style: normal;
-  color: var(--sapContent_NonInteractiveIconColor);
-  background-color: var(--sapBackgroundColor);
-  display: inline-block;
-
-  @each $key, $value in $fd-icons {
-    &--#{$key} {
-      @include fd-icon-glyph($key);
-    }
-  }
-}
