+++ src/fundamental-styles.scss
@@ -17,7 +17,6 @@
 @import "./checkbox";
 @import "./counter";
 @import "./dialog";
-@import "./fd-icon";
 @import "./fieldset";
 @import "./form-layout-grid";
 @import "./form-group";
