+++ src/icon.scss
@@ -1,16 +1,17 @@
 @import "./mixins";
 
-$block: sap-icon;
+$block: #{$fd-namespace}-icon;
 
-[class*="#{$block}"] {
+.#{$block} {
   @include fd-reset();
   @include fd-icon-base();
 
+  font-size: 1rem;
   font-style: normal;
+  color: var(--sapContent_NonInteractiveIconColor);
+  background-color: var(--sapBackgroundColor);
   display: inline-block;
-}
 
-.#{$block} {
   @each $key, $value in $fd-icons {
     &--#{$key} {
       @include fd-icon-glyph($key);
